requirements.txt
asgiref==3.7.2
cleanup==1.0.4
Django==4.2.8
django-cleanup==8.1.0
djangorestframework==3.14.0
docopt==0.6.2
huepy==1.2.1
pillow==10.2.0
pytz==2024.1
sqlparse==0.4.4
tzdata==2024.1
